NEO Blockchain scrapper int to float64

+++ blockchain-scrapers/cmd/neo/neo.go
@@ -85,7 +85,7 @@ func main() {
 
 					client.SendSupply(&dia.Supply{
 						Symbol:            "NEO",
-						CirculatingSupply: circulatingSupply,
+						CirculatingSupply: float64(circulatingSupply),
 					})
 
 					lastCirculatedSupply = circulatingSupply
